+++ src/components/Home/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import Banner from './Banner';
+import Banner from './Banner/Banner';
 import MainView from './MainView';
 import Tags from './Tags';
 import agent from '../../agent';
