+++ src/reducers/common.js
@@ -18,7 +18,7 @@ import {
 } from '../constants/actionTypes';
 
 const defaultState = {
-  appName: 'Practicum Project Kitchen',
+  appName: 'Проектная кухня',
   token: null,
   viewChangeCounter: 0,
 };
